compat: do not expose chunk_compat.h as part of library interface
Previously, chunkio exposed functions and macros in chunk_compat.h
as part of library interface. This caused a number of issues in the
downstream source code (e.g. conflicts of macro definitions).

This patch prevents the compatibility header from leaking to the
outer world, while keeping the source tree compilable.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>