file: fix 'file' size adjustment, always honor requested new size
The function in charge to adjust the size of the file into the
file system, was doing a size comparisson before to perform the
adjustment. The comparisson was wrong since it assumed the 'cf'
context has the real 'file system' file size (st.st_size).

This patch removes the comparisson; next patches adds a test
case to avoid regressions.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>