stream: new API to retrieve total stream 'up' size
On high load environment with thousands of chunks, it takes time
to count the number of bytes being used by the active chunks.

This patch implements a new mechanism where Chunks are now linked
in an additional linked list for 'up' or 'down' chunks. Through
a new API that iterate only the 'up' chunks now is possible to
retrieve the bytes size in a direct way.

- Problematic old case before the patch -

  - 40k chunks in the file system
  - 128 chunks are up

Counting the number of bytes needed to iterate the 40k entries.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>