file: set fs_size if file cannot be mapped due config restrictions
If a pre-existent chunk file cannot be opened properly 'up' in memory
due to configuration restriction like max chunks up, the code was
returning the context but not setting up the real size of the file
in the structure. The caller of Chunk I/O library when trying to
get the file size by using cio_chunk_get_real_size() is always receiving
a '0' value.

This patch enforce a file size check even if the file is not up in
memory.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>