file: fix checksum handling and close file descriptor
If 'checksum' option has been enabled and the caller used the function
cio_chunk_write_at(), the 'current' checksum  will likely be corrupted,
since the caller might want to write data to a previous offset (lowest
position) from the last 'current crc' that was calculated, the function
cio_chunk_write_at() is not updating the checksums, so if the chunk was
set down() and up() again, the corruption gets visible.

As a side effect, if the user wants to put the corrupted chunk with
'bad checksum up', it was leaving an open file descriptor to the chunk.

The following patch fixes the issues mentioned above:

- when invoking cio_chunk_write_at() the checksums are now
  updated.

- when opening up a corrupted chunk, if it's because a bad checksum
  the file descriptor is now closed.

In addition this patch uses the new 'error' interface to set the
optional error number in the chunk, so the caller can inspect more
about a possible corruption or error.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>