file: fix fallback mechanism for fallocate(2) on Linux
When changing the file size in Linux when fallocate(2) is present,
the current implementation has a problem when checking 'why' the
syscall failed, it's wrongly checking the return code instead of the
errno variable.

This patch fix the return code and errno checks for EOPNOTSUPP. Despite
the running system can support fallocate(2) does not means memory mapped
files are supported, this happen in older kernels.

Now when EOPNOTSUPP happens with fallocate(2), there is a fallback mechanism
through posix_fallocate() and that fallback choice is preserved in the file
context for further operations.

Signed-off-by: Eduardo Silva <eduardo@calyptia.com>