file: increase write buffer size and remove truncation
When writing data to a chunk, the file backend allocate extra space
as needed and uses a factor of 32KB (realloc_size). For concurrent
writes this incurs in many fallocate() syscalls which have a performance
penalty. By increasing the size of the write buffer to 512KB we reduce
the number of system calls needed, as a side effect it becomes more
performant.

In addition, when Chunks are put in a 'down' state, those are truncated
because most of the time they have extra empty space. In a saturated
environment where the number of Chunks up in memory are hitting the
limit of max_chunks_up, the system does a lot of up/dows of Chunks, again
when truncating files falllocate() is invoked. By not truncating the
files improves performance.

Note that the only issue with not truncating the files is that checksums
will always fails, so it's expected to have checksums disabled.

Signed-off-by: Eduardo Silva <eduardo@calyptia.com>